third_party/WebKit/LayoutTests/fullscreen/full-screen-remove-expected.txt

EVENT(webkitfullscreenchange)
EXPECTED (document.webkitCurrentFullScreenElement == '[object HTMLSpanElement]') OK
EVENT(webkitfullscreenchange)
EXPECTED (document.webkitCurrentFullScreenElement == 'any') OK
END OF TEST
